tests: refactor test structure.
