+++ test/unit/command/orders.command.mock.ts
@@ -1,4 +1,4 @@
-import { signAndSend } from "../@polkadot-api.mock.ts";
+import { signAndSend } from "../@polkadot-api.mock";
 
 export const orders = {
   createOrder(param) {
