+++ test/unit/query/user-profile/user-profile.query.spec.ts
@@ -1,9 +1,9 @@
-import { queryAccountIdByEthAddress, queryEthAdressByAccountId } from "../../../src/query/user-profile";
-import { ApiPromise } from "../@polkadot-api.mock.ts";
-import { mockFunction } from "../mock";
+import { queryAccountIdByEthAddress, queryEthAdressByAccountId } from "../../../../src/query/user-profile";
+import { ApiPromise } from "../../@polkadot-api.mock";
+import { mockFunction } from "../../mock";
 import { userProfile } from "./user-profile.query.mock";
 
-jest.mock('../mock', () => ({
+jest.mock('../../mock', () => ({
   mockFunction: jest.fn(),
 }));
